requirements.txt
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
colorama==0.4.6
distro==1.9.0
exceptiongroup==1.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.13.0
keyboard==0.13.5
openai==2.21.0
pillow==12.1.1
pydantic==2.12.5
pydantic_core==2.41.5
pyperclip==1.11.0
pystray==0.19.5
six==1.17.0
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
